requirements.txt
beautifulsoup4==4.12.2
python-dotenv==1.0.1
requests==2.31.0